// ui/src/generated/MediaPage.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { MediaSummary } from "./MediaSummary";
import type { OperationId } from "./OperationId";

export type MediaPage = { operation_id: OperationId, media: Array<MediaSummary>, next_sequence: number, has_more: boolean, gap: boolean, };